Polish DESCRIPTION
DESCRIPTION is a picky file. It is a very special file because it
defines an R package. CRAN is very picky about how to write it;
for example, the title must use Title Case not Sentense case and
not end in dot ".". Also, they try to avoid the use of the word
'package' and they ask for package names.

This convensions are totally idiosyncratic and not worth learning
unless you submit to CRAN. But as I'm here, why not try meet those
CRAN policies?