Avoid importing needless functions
What we use from r2dii.data is currently accessed via the syntax
"r2dii.data::objectname". We could do it with the
sytax  "#' importFrom r2dii.data objectname" but this syntax
is better avoided "#' @import r2dii.data" because it brings all
functions including the ones we don't use and increases the
chances for conflicts. For more see the book R packages.