Compare with r2dii.analysis::target_market_share()
The commit title avoids refering to 'banks data' because it is
ambiguous. I prefer to be more specific and say that the output
of get_example_data() is expected to have the same structure than
the output of r2dii.analysis::target_market_share().

I think this more precise wording should also be used in the
helpfile.