plot_trajectoryX() looses arg 'value' for consistency with X API (#231)
`plot_trajectoryX()` looses the argument 'value' for consistency with X API

This is the opposite of what we discussed with Monika in our meeting, but
it makes makes more sense. When we discussed we didn't notice that the
X (thin) API does not have `main` or `value` -- only the Y (thick) API.
This is consistent with A/B testing two extreemes: thin and thick.