This is a combination of 2 commits.
New minimal ggplot_* leave plot_* at high-level

Closes #355

This PR needs careful review as it implements an idea we have not
yet discussed.

--

The plot_*() funciton again beautify labels.
New ggplot*() functions do not. They are intended as lower level
function that we can use internally or export them (not yet) to
give user the hability to customize their plots themselves.

That is, the plot_*() functions are here what we discussed it would
become the qplot_*() functions. And ggplot_*() functions are here
what we discussed it would be the plot_*() functions.