Polish documentation
* Remove `@noRd` where it's not necessary. This tag stops roxygen2
from creating an R-documentation (.Rd) file, so we use to write
developer-facing documentation that should not result in an .Rd
file (the ones we access with `?functionname`). But if a function
has no documentation at all (most, simple internal functions don't)
then roxygen won't create an .Rd file anyway, so @noRd adds nothing.

* New family 'r2dii scales' generate a link between all `scale_*()`
functions -- i.e. the section "See also". This helps users navigate
the documentation of similar functions. It also helps create the
reference section of the website in a more specific way: i.e. with
`has_concept('name of the family')` instead of
`starts_with('prefix')`.