Separate data from plot
The motivation to avoid a confusing warning when the data is piped
into qplot_techmix() rather than passed as the first argument. In
that case the warning prints:

   Do you want to plot different years? E.g. filter . with:`subset(., year %in% c(2020, 2030))`.

Note the "." where you expect "data".

Along the way, I noted inconsistent use of pipeing versus passing
data, so here I ensure all calls pass data.

I also like this because it makes a clearer distinction between
manipulation of data and plots.